# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Book API",
        "description": "Мікросервісне API для управління базою книг",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Список книг",
                "operationId": "921c595372ab399821e267fcf05cbaa1",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успішно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/BookList"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Створення книги",
                "operationId": "7ab92284110ace5d0f7de52f372e70c0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BookStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Книгу створено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/BookDetail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Помилка валідації"
                    }
                }
            }
        },
        "/books/{id}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Детальна інформація про книгу",
                "operationId": "e079e587c1fee53fe150984526abf3ef",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успішно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/BookDetail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Книгу не знайдено"
                    }
                }
            },
            "put": {
                "tags": [
                    "Books"
                ],
                "summary": "Оновлення книги",
                "operationId": "b47b1606570ba86a105bf62f04de6b7f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BookStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Книгу оновлено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/BookDetail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Книгу не знайдено"
                    },
                    "422": {
                        "description": "Помилка валідації"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Books"
                ],
                "summary": "Видалення книги",
                "operationId": "95c7fc7d78f9dc3884a5bb898b7ead11",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Книгу видалено"
                    },
                    "404": {
                        "description": "Книгу не знайдено"
                    }
                }
            }
        },
        "/books/import": {
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Імпорт книг з CSV",
                "operationId": "4f9ac10694055258e52050a0f8ed7912",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "CSV файл (максимум 10 МБ)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Файл прийнято, імпорт виконується асинхронно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "queued"
                                        },
                                        "path": {
                                            "type": "string",
                                            "example": "imports/books.csv"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Помилка валідації"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "BookList": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "The Great Gatsby"
                    },
                    "authors": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "F. Scott Fitzgerald"
                        ]
                    },
                    "publisher": {
                        "type": "string",
                        "example": "Scribner",
                        "nullable": true
                    },
                    "year": {
                        "type": "string",
                        "example": "1925",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "BookDetail": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/BookList"
                    },
                    {
                        "properties": {
                            "description": {
                                "type": "string",
                                "nullable": true
                            },
                            "edition": {
                                "type": "integer",
                                "example": 1,
                                "nullable": true
                            },
                            "pages": {
                                "type": "integer",
                                "example": 180,
                                "nullable": true
                            },
                            "format": {
                                "type": "string",
                                "example": "Hardcover",
                                "nullable": true
                            },
                            "country": {
                                "type": "string",
                                "example": "USA",
                                "nullable": true
                            },
                            "isbn": {
                                "type": "string",
                                "example": "9780743273565",
                                "nullable": true
                            },
                            "genres": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                },
                                "example": [
                                    "Novel",
                                    "Fiction"
                                ]
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "BookStoreRequest": {
                "required": [
                    "title"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "The Great Gatsby"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "edition": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "isbn": {
                        "type": "string",
                        "example": "9780743273565",
                        "nullable": true
                    },
                    "published_date": {
                        "type": "string",
                        "format": "date",
                        "example": "1925-04-10",
                        "nullable": true
                    },
                    "pages": {
                        "type": "integer",
                        "example": 180,
                        "nullable": true
                    },
                    "format": {
                        "type": "string",
                        "example": "Hardcover",
                        "nullable": true
                    },
                    "country": {
                        "type": "string",
                        "example": "USA",
                        "nullable": true
                    },
                    "publisher_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "author_ids": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            1,
                            2
                        ],
                        "nullable": true
                    },
                    "genre_ids": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            1,
                            2
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 5
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "total": {
                        "type": "integer",
                        "example": 50
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Books",
            "description": "Books"
        }
    ]
}